Refactor code for linter errors
