--- src/removeTasks.js
@@ -1,12 +1,0 @@
-import { tasks } from './index.js';
-
-export const ul = document.querySelector('.list');
-
-export default function removeTasks(event) {
-  event.stopPropagation();
-  const li = event.target.parentNode;
-  const index = Array.prototype.indexOf.call(ul.children, li);
-  tasks.splice(index, 1);
-  li.remove();
-  localStorage.setItem('tasks', JSON.stringify(tasks));
-}
